Fix potential memory leak in update_metadata
